requirements.txt
flask
flask-cors
pandas
requests
google-genai
google-adk
matplotlib
gunicorn
